Refactor transform code
